+++ src/util.cpp
@@ -384,17 +384,17 @@ bool chimera::util::needsReturnValuePolicy(const NamedDecl *decl, const Type *re
     return false;
 }
 
-std::pair<size_t, size_t>
+std::pair<unsigned, unsigned>
 chimera::util::getFunctionArgumentRange(const clang::FunctionDecl *decl)
 {
-    const size_t max_arguments = decl->getNumParams();
+    const unsigned max_arguments = decl->getNumParams();
 
-    size_t min_arguments = 0;
+    unsigned min_arguments = 0;
     for (; min_arguments < max_arguments; ++min_arguments)
     {
         if (decl->getParamDecl(min_arguments)->hasDefaultArg())
             break;
     }
 
-    return std::pair<size_t, size_t>(min_arguments, max_arguments);
+    return std::pair<unsigned, unsigned>(min_arguments, max_arguments);
 }
